Rename imports; replace old usages
