+++ python/semantic_kernel/semantic_functions/prompt_template.py
@@ -10,8 +10,8 @@ from semantic_kernel.semantic_functions.prompt_template_config import (
 from semantic_kernel.skill_definition.parameter_view import ParameterView
 from semantic_kernel.template_engine.blocks.block_types import BlockTypes
 from semantic_kernel.template_engine.blocks.var_block import VarBlock
-from semantic_kernel.template_engine.prompt_template_engine_base import (
-    PromptTemplateEngineBase,
+from semantic_kernel.template_engine.protocols.prompt_templating_engine import (
+    PromptTemplatingEngine,
 )
 from semantic_kernel.utils.null_logger import NullLogger
 
@@ -21,14 +21,14 @@ if TYPE_CHECKING:
 
 class PromptTemplate(PromptTemplateBase):
     _template: str
-    _template_engine: PromptTemplateEngineBase
+    _template_engine: PromptTemplatingEngine
     _log: Logger
     _prompt_config: PromptTemplateConfig
 
     def __init__(
         self,
         template: str,
-        template_engine: PromptTemplateEngineBase,
+        template_engine: PromptTemplatingEngine,
         prompt_config: PromptTemplateConfig,
         log: Optional[Logger] = None,
     ) -> None:
