+++ python/semantic_kernel/template_engine/blocks/block.py
@@ -3,7 +3,7 @@
 from logging import Logger
 from typing import Optional, Tuple
 
-from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine.blocks.block_types import BlockTypes
 from semantic_kernel.utils.null_logger import NullLogger
