+++ python/semantic_kernel/template_engine/blocks/code_block.py
@@ -7,11 +7,11 @@ from semantic_kernel.orchestration.sk_function_base import SKFunctionBase
 from semantic_kernel.skill_definition.read_only_skill_collection_base import (
     ReadOnlySkillCollectionBase,
 )
-from semantic_kernel.template_engine_v2.blocks.block import Block
-from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
-from semantic_kernel.template_engine_v2.blocks.function_id_block import FunctionIdBlock
-from semantic_kernel.template_engine_v2.code_tokenizer import CodeTokenizer
-from semantic_kernel.template_engine_v2.protocols.code_renderer import CodeRenderer
+from semantic_kernel.template_engine.blocks.block import Block
+from semantic_kernel.template_engine.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine.blocks.function_id_block import FunctionIdBlock
+from semantic_kernel.template_engine.code_tokenizer import CodeTokenizer
+from semantic_kernel.template_engine.protocols.code_renderer import CodeRenderer
 
 
 class CodeBlock(Block, CodeRenderer):
