+++ python/semantic_kernel/template_engine/blocks/text_block.py
@@ -4,9 +4,9 @@ from logging import Logger
 from typing import Optional, Tuple
 
 from semantic_kernel.orchestration.context_variables import ContextVariables
-from semantic_kernel.template_engine_v2.blocks.block import Block
-from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
-from semantic_kernel.template_engine_v2.protocols.text_renderer import TextRenderer
+from semantic_kernel.template_engine.blocks.block import Block
+from semantic_kernel.template_engine.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine.protocols.text_renderer import TextRenderer
 
 
 class TextBlock(Block, TextRenderer):
