+++ python/semantic_kernel/template_engine/prompt_template_engine.py
@@ -3,15 +3,15 @@
 from typing import Optional, List
 from logging import Logger
 
-from semantic_kernel.template_engine_v2.blocks.block import Block
-from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
-from semantic_kernel.template_engine_v2.blocks.text_block import TextBlock
-from semantic_kernel.template_engine_v2.protocols.prompt_templating_engine import (
+from semantic_kernel.template_engine.blocks.block import Block
+from semantic_kernel.template_engine.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine.blocks.text_block import TextBlock
+from semantic_kernel.template_engine.protocols.prompt_templating_engine import (
     PromptTemplatingEngine,
 )
-from semantic_kernel.template_engine_v2.protocols.text_renderer import TextRenderer
-from semantic_kernel.template_engine_v2.protocols.code_renderer import CodeRenderer
-from semantic_kernel.template_engine_v2.template_tokenizer import TemplateTokenizer
+from semantic_kernel.template_engine.protocols.text_renderer import TextRenderer
+from semantic_kernel.template_engine.protocols.code_renderer import CodeRenderer
+from semantic_kernel.template_engine.template_tokenizer import TemplateTokenizer
 from semantic_kernel.orchestration.sk_context import SKContext
 from semantic_kernel.orchestration.context_variables import ContextVariables
 from semantic_kernel.utils.null_logger import NullLogger
