+++ python/semantic_kernel/template_engine/protocols/prompt_templating_engine.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Protocol
 
 from semantic_kernel.orchestration.context_variables import ContextVariables
 from semantic_kernel.orchestration.sk_context import SKContext
-from semantic_kernel.template_engine_v2.blocks.block import Block
+from semantic_kernel.template_engine.blocks.block import Block
 
 
 class PromptTemplatingEngine(Protocol):
