+++ python/tests/template_engine/blocks/test_block.py
@@ -4,8 +4,8 @@ from logging import Logger
 
 from pytest import raises
 
-from semantic_kernel.template_engine_v2.blocks.block import Block
-from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine.blocks.block import Block
+from semantic_kernel.template_engine.blocks.block_types import BlockTypes
 from semantic_kernel.utils.null_logger import NullLogger
