+++ python/tests/template_engine/blocks/test_function_id_block.py
@@ -5,8 +5,8 @@ from logging import Logger
 from pytest import mark, raises
 
 from semantic_kernel.orchestration.context_variables import ContextVariables
-from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
-from semantic_kernel.template_engine_v2.blocks.function_id_block import FunctionIdBlock
+from semantic_kernel.template_engine.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine.blocks.function_id_block import FunctionIdBlock
 
 
 def test_init():
