+++ python/tests/template_engine/blocks/test_val_block.py
@@ -3,8 +3,8 @@
 from logging import Logger
 
 from semantic_kernel.orchestration.context_variables import ContextVariables
-from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
-from semantic_kernel.template_engine_v2.blocks.val_block import ValBlock
+from semantic_kernel.template_engine.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine.blocks.val_block import ValBlock
 
 
 def test_init():
