+++ python/tests/template_engine/test_code_tokenizer.py
@@ -3,8 +3,8 @@
 from logging import Logger
 from pytest import raises, mark
 
-from semantic_kernel.template_engine_v2.blocks.block_types import BlockTypes
-from semantic_kernel.template_engine_v2.code_tokenizer import CodeTokenizer
+from semantic_kernel.template_engine.blocks.block_types import BlockTypes
+from semantic_kernel.template_engine.code_tokenizer import CodeTokenizer
 
 
 def test_it_parses_empty_text():
